+++ src/main/java/com/gmail/val59000mc/scenarios/scenariolisteners/BloodDiamondsListener.java
@@ -1,9 +1,8 @@
 package com.gmail.val59000mc.scenarios.scenariolisteners;
 
+import com.gmail.val59000mc.players.UhcPlayer;
 import com.gmail.val59000mc.scenarios.ScenarioListener;
 import com.gmail.val59000mc.utils.OreType;
-import com.gmail.val59000mc.utils.UniversalSound;
-import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.block.BlockBreakEvent;
@@ -16,14 +15,7 @@ public class BloodDiamondsListener extends ScenarioListener{
 			return;
 		}
 
-		Player p = e.getPlayer();
-		p.getWorld().playSound(p.getLocation(), UniversalSound.PLAYER_HURT.getSound(), 1, 1);
-
-		if (p.getHealth() < 1){
-			p.setHealth(0);
-		}else {
-			p.setHealth(p.getHealth() - 1);
-		}
+		UhcPlayer.damageIrreducible(e.getPlayer(), 1);
 	}
 
 }